+++ dao/fake.go
@@ -79,6 +79,11 @@ func (fake *FakeDAO) DeleteObjectType(objectType models.ODObjectType) error {
 	return fake.Err
 }
 
+// Expunge objects deleted by user.
+func (fake *FakeDAO) ExpungeDeletedByUser(user models.ODUser, pageSize int) (int64, error) {
+	return int64(0), fake.Err
+}
+
 // ExpungeObject for FakeDAO.
 func (fake *FakeDAO) ExpungeObject(user models.ODUser, object models.ODObject, explicit bool) error {
 	return fake.Err